장바구니 삭제할때 상품명 보여주면서 한번 더 컨펌, 할인안된 상품 금액 추가

+++ src/main/resources/static/js/shoppingBag.js
@@ -18,7 +18,7 @@ document.addEventListener('DOMContentLoaded', function () {
     const allChoose = document.getElementById('All-choose');
     const itemCheckboxes = document.querySelectorAll('.item-checkbox');
     const productPrices = document.querySelectorAll('.product-sprice');
-    const productoPrices=document.querySelectorAll('.product-oprice');
+    const productoPrices = document.querySelectorAll('.product-oprice');
     const deliveryFeeElements = document.querySelectorAll('.delivery-fee');
     const totalPriceElement = document.getElementById('total-price');
     const deliveryFeeElement = document.getElementById('delivery-fee');
@@ -107,7 +107,7 @@ document.addEventListener('DOMContentLoaded', function () {
                 totalOprice += oprice * quantity;
                 totalSprice += sprice * quantity;
 
-                console.log("total: " + totalOprice + "  "+totalSprice);
+                console.log("total: " + totalOprice + "  " + totalSprice);
             }
         });
         console.log(`Total oprice: ${totalOprice}`);
@@ -116,19 +116,24 @@ document.addEventListener('DOMContentLoaded', function () {
         // 총 상품 금액 및 배송비 업데이트
         totalPriceElement.textContent = totalPrice.toLocaleString() + '원';
         deliveryFeeElement.textContent = totalDeliveryFee.toLocaleString() + '원';
-        document.getElementById("total-price").value=totalPrice;
-        document.getElementById("delivery-fee").value=totalDeliveryFee;
+        document.getElementById("total-price").value = totalPrice;
+        document.getElementById("delivery-fee").value = totalDeliveryFee;
+        document.getElementById("total-oprice").value = totalOprice;
+        console.log("뿍빡", totalOprice);
 
         // 결제 금액 업데이트
         const finalPrice = totalPrice + totalDeliveryFee;
         finalPriceElement.textContent = finalPrice.toLocaleString() + '원';
-        document.getElementById('final-price').value=finalPrice;
+        document.getElementById('final-price').value = finalPrice;
 
         // oprice와 sprice의 총합 차액 계산
         const totalSale = totalOprice - totalSprice;
         totalSaleElement.textContent = totalSale.toLocaleString() + '원';
-        document.getElementById("total-sale").value=totalSale;
-        console.log("세일된 금액 : " ,totalSale);
+        document.getElementById("total-sale").value = totalSale;
+        console.log("세일된 금액 : ", totalSale);
+
+        const what = document.getElementById("total-price").value;
+        console.log("total_price: ", what);
     }
 
     // 초기화
@@ -149,9 +154,6 @@ document.addEventListener('DOMContentLoaded', function () {
     });
 
 
-
-
-
     // 결제하기 버튼 클릭 이벤트
     const submitButton = document.getElementById('submit_button');
     submitButton.addEventListener('click', function (event) {
@@ -166,30 +168,28 @@ document.addEventListener('DOMContentLoaded', function () {
             const productNum = document.getElementById(`product_num[${index}]`).value;
             const cartCnt = document.getElementById(`cart_cnt[${index}]`).value;
             const userNum = document.getElementById('user_num').value;
-            const totalOprice=document.getElementById('total-price').value;
-            const totalSale=document.getElementById('total-sale').value;
-            const deliveryFee=document.getElementById('delivery-fee').value;
-            const finalPrice=document.getElementById('final-price').value;
+            const totalOprice = document.getElementById('total-oprice').value;
+            const totalSale = document.getElementById('total-sale').value;
+            const deliveryFee = document.getElementById('delivery-fee').value;
+            const finalPrice = document.getElementById('final-price').value;
 
             orderData.push({
                 cart_num: cartNum,
                 product_num: productNum,
                 cart_cnt: cartCnt,
                 user_num: userNum,
-                total_oprice : totalOprice,
-                total_sale:totalSale,
-                total_delivery:deliveryFee,
-                total_price : finalPrice
+                total_oprice: totalOprice,
+                total_sale: totalSale,
+                total_delivery: deliveryFee,
+                total_price: finalPrice
             });
         });
 
         // 서버에 데이터 전송
         fetch('/my/payment', {
-            method: 'POST',
-            headers: {
+            method: 'POST', headers: {
                 'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(orderData)
+            }, body: JSON.stringify(orderData)
         })
             .then(response => response.text())
             .then(data => {
@@ -206,30 +206,34 @@ document.addEventListener('DOMContentLoaded', function () {
 
 //장바구니 상품 삭제
 function cartDelete(index) {
-    const product_num = document.getElementById(`product_num[${index}]`).value;
-    const user_num = document.getElementById("user_num").value;
-    console.log(user_num, product_num);
+    const product_name = document.getElementById(`product_name[${index}]`).value;
 
-    fetch('/cart/delete', {
-        method: 'DELETE', headers: {
-            'Content-Type': 'application/json'
-        }, body: JSON.stringify({
-            product_num: product_num, user_num: user_num
-        })
-    }).then(response => {
-        if (response.ok) {
-            // 요청이 성공했을 경우, DOM에서 해당 항목을 삭제
-            const cartItem = document.getElementById(`cart-item[${index}]`);
-            if (cartItem) {
-                cartItem.remove();
+    if (confirm(product_name + "을(를) 삭제하시겠습니까?")) {
+        const product_num = document.getElementById(`product_num[${index}]`).value;
+        const user_num = document.getElementById("user_num").value;
+        console.log(user_num, product_num);
+
+        fetch('/cart/delete', {
+            method: 'DELETE', headers: {
+                'Content-Type': 'application/json'
+            }, body: JSON.stringify({
+                product_num: product_num, user_num: user_num
+            })
+        }).then(response => {
+            if (response.ok) {
+                // 요청이 성공했을 경우, DOM에서 해당 항목을 삭제
+                const cartItem = document.getElementById(`cart-item[${index}]`);
+                if (cartItem) {
+                    cartItem.remove();
+                }
+                console.log("삭제 완료:", "user_num:", user_num, "product_num:", product_num);
+            } else {
+                console.error('삭제 실패');
             }
-            console.log("삭제 완료:", "user_num:", user_num, "product_num:", product_num);
-        } else {
-            console.error('삭제 실패');
-        }
-    }).catch(error => {
-        console.error('요청 중 오류 발생:', error);
-    });
+        }).catch(error => {
+            console.error('요청 중 오류 발생:', error);
+        });
+    }
 }
 
 //장바구니 상품 수정(수량 수정)
@@ -252,4 +256,11 @@ function cartUpdate(index) {
             cart_cnt: cart_cnt
         })
     }).then(response => response.text())
+        .then(data => {
+            alert("수량이 성공적으로 변경되었습니다!");
+            console.log(data); // 서버 응답 데이터 확인용
+        }).catch(error => {
+            console.error("에러 발생: ", error);
+            alert("수량 변경에 실패했습니다. 다시 시도해주세요.");
+        });
 }
